Move and clean up unnecessary files
