--- frontend/src/components/Search/Parser/.antlr/SearchGrammar.interp
@@ -1,46 +1,0 @@
-token literal names:
-null
-'AND'
-'OR'
-'NOT'
-'='
-'!='
-'<'
-'<='
-'>'
-'>='
-'('
-')'
-':'
-null
-null
-null
-
-token symbolic names:
-null
-AND
-OR
-NOT
-EQ
-NEQ
-LT
-LTE
-GT
-GTE
-LPAREN
-RPAREN
-COLON
-ID
-STRING
-WS
-
-rule names:
-search_query
-col_expr
-search_expr
-search_key
-bin_op
-
-
-atn:
-[4, 1, 15, 74, 2, 0, 7, 0, 2, 1, 7, 1, 2, 2, 7, 2, 2, 3, 7, 3, 2, 4, 7, 4, 1, 0, 1, 0, 1, 0, 1, 0, 3, 0, 15, 8, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 3, 1, 26, 8, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 5, 1, 34, 8, 1, 10, 1, 12, 1, 37, 9, 1, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 3, 2, 55, 8, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 5, 2, 65, 8, 2, 10, 2, 12, 2, 68, 9, 2, 1, 3, 1, 3, 1, 4, 1, 4, 1, 4, 0, 2, 2, 4, 5, 0, 2, 4, 6, 8, 0, 1, 2, 0, 4, 9, 12, 12, 81, 0, 14, 1, 0, 0, 0, 2, 25, 1, 0, 0, 0, 4, 54, 1, 0, 0, 0, 6, 69, 1, 0, 0, 0, 8, 71, 1, 0, 0, 0, 10, 15, 5, 0, 0, 1, 11, 12, 3, 4, 2, 0, 12, 13, 5, 0, 0, 1, 13, 15, 1, 0, 0, 0, 14, 10, 1, 0, 0, 0, 14, 11, 1, 0, 0, 0, 15, 1, 1, 0, 0, 0, 16, 17, 6, 1, -1, 0, 17, 18, 5, 10, 0, 0, 18, 19, 3, 2, 1, 0, 19, 20, 5, 11, 0, 0, 20, 26, 1, 0, 0, 0, 21, 22, 5, 3, 0, 0, 22, 26, 3, 2, 1, 3, 23, 26, 5, 14, 0, 0, 24, 26, 5, 13, 0, 0, 25, 16, 1, 0, 0, 0, 25, 21, 1, 0, 0, 0, 25, 23, 1, 0, 0, 0, 25, 24, 1, 0, 0, 0, 26, 35, 1, 0, 0, 0, 27, 28, 10, 5, 0, 0, 28, 29, 5, 1, 0, 0, 29, 34, 3, 2, 1, 6, 30, 31, 10, 4, 0, 0, 31, 32, 5, 2, 0, 0, 32, 34, 3, 2, 1, 5, 33, 27, 1, 0, 0, 0, 33, 30, 1, 0, 0, 0, 34, 37, 1, 0, 0, 0, 35, 33, 1, 0, 0, 0, 35, 36, 1, 0, 0, 0, 36, 3, 1, 0, 0, 0, 37, 35, 1, 0, 0, 0, 38, 39, 6, 2, -1, 0, 39, 40, 5, 10, 0, 0, 40, 41, 3, 4, 2, 0, 41, 42, 5, 11, 0, 0, 42, 55, 1, 0, 0, 0, 43, 44, 5, 10, 0, 0, 44, 45, 3, 4, 2, 0, 45, 46, 5, 11, 0, 0, 46, 55, 1, 0, 0, 0, 47, 48, 5, 3, 0, 0, 48, 55, 3, 4, 2, 3, 49, 50, 3, 6, 3, 0, 50, 51, 3, 8, 4, 0, 51, 52, 3, 2, 1, 0, 52, 55, 1, 0, 0, 0, 53, 55, 3, 2, 1, 0, 54, 38, 1, 0, 0, 0, 54, 43, 1, 0, 0, 0, 54, 47, 1, 0, 0, 0, 54, 49, 1, 0, 0, 0, 54, 53, 1, 0, 0, 0, 55, 66, 1, 0, 0, 0, 56, 57, 10, 6, 0, 0, 57, 58, 5, 1, 0, 0, 58, 65, 3, 4, 2, 7, 59, 60, 10, 5, 0, 0, 60, 65, 3, 4, 2, 6, 61, 62, 10, 4, 0, 0, 62, 63, 5, 2, 0, 0, 63, 65, 3, 4, 2, 5, 64, 56, 1, 0, 0, 0, 64, 59, 1, 0, 0, 0, 64, 61, 1, 0, 0, 0, 65, 68, 1, 0, 0, 0, 66, 64, 1, 0, 0, 0, 66, 67, 1, 0, 0, 0, 67, 5, 1, 0, 0, 0, 68, 66, 1, 0, 0, 0, 69, 70, 5, 13, 0, 0, 70, 7, 1, 0, 0, 0, 71, 72, 7, 0, 0, 0, 72, 9, 1, 0, 0, 0, 7, 14, 25, 33, 35, 54, 64, 66]
