--- frontend/src/components/Search/Parser/.antlr/SearchGrammar.tokens
@@ -1,27 +1,0 @@
-AND=1
-OR=2
-NOT=3
-EQ=4
-NEQ=5
-LT=6
-LTE=7
-GT=8
-GTE=9
-LPAREN=10
-RPAREN=11
-COLON=12
-ID=13
-STRING=14
-WS=15
-'AND'=1
-'OR'=2
-'NOT'=3
-'='=4
-'!='=5
-'<'=6
-'<='=7
-'>'=8
-'>='=9
-'('=10
-')'=11
-':'=12
